Get unit tests to compile
