+++ bin/prover-perf/Cargo.toml
@@ -39,5 +39,5 @@ tracing.workspace = true
 
 [features]
 default = ["sp1"]
-sp1 = ["zkaleido-sp1-host/perf", "strata-zkvm-hosts/sp1-builder"]
 risc0 = ["zkaleido-risc0-host/perf", "strata-zkvm-hosts/risc0-builder"]
+sp1 = ["zkaleido-sp1-host/perf", "strata-zkvm-hosts/sp1-builder"]
