+++ crates/l1tx/src/lib.rs
@@ -3,3 +3,5 @@ pub mod envelope;
 pub mod filter;
 pub mod messages;
 pub mod utils;
+
+pub use filter::types::TxFilterConfig;
